requirements.txt
alembic==1.4.3
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.2.0
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
distlib==0.3.1
fastapi==0.63.0
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.0.0
h11==0.12.0
identify==2.2.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
loguru==0.5.3
Mako==1.1.4
MarkupSafe==1.1.1
nodeenv==1.5.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pluggy==0.13.1
pre-commit==2.11.1
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.3
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
virtualenv==20.4.3
Werkzeug==1.0.1
